Moved route provisionning to Vagrantfile as it is more consistent
It seems that the way vagrant handle networking makes it difficult to
set up static routes init scripts. Having it in the Vagrantfile as an
run always provisionner ensure tht it is run after network setup when
doing vagrant up. always start the VM with vagrant up.
Also added a Windows .cmd to setup route in an elevated command prompt.