Added sdkman, maven, gradle. add private bin
Install sdkman, use it to install maven and gradle.
Add ~/bin which will be added to PATH.
Force LF for *.sh to make sure they work on Windows host after clone
when provisionning VM.
Added a few snippets to README.md in labs/
(this is quite many different things for just one commit).